requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alpaca-trade-api==3.2.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
frozenlist==1.5.0
idna==3.10
msgpack==1.0.3
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
propcache==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3
